Fix for stepped loops
Preliminary fix for stepped loops where the iteration won't
hit the upper bound directly, just pass over it.
This fix “works” on the compiler I tried but it probably isn’t
portable since it relies on the other of arguments to
operator== in the language implementation of the range-based
for loop (which is implied in §6.5.4).